[IR] Don't include PassInstrumentation.h in PassManager.h (NFC)
Move PassInstrumentationAnalysis into PassInstrumentation.h and
stop including it in PassManager.h (effectively inverting the
direction of the dependency).

Most places using PassManager are not interested in
PassInstrumentation, and we no longer have any uses of it in
PassManager.h itself (only in PassManagerImpl.h).
