+++ llvm/include/llvm/IR/PassManagerImpl.h
@@ -16,6 +16,7 @@
 #define LLVM_IR_PASSMANAGERIMPL_H
 
 #include "llvm/Support/CommandLine.h"
+#include "llvm/IR/PassInstrumentation.h"
 #include "llvm/IR/PassManager.h"
 
 extern llvm::cl::opt<bool> UseNewDbgInfoFormat;
