+++ llvm/include/llvm/Transforms/Utils/Debugify.h
@@ -20,6 +20,7 @@
 #include "llvm/IR/IRPrintingPasses.h"
 #include "llvm/IR/LegacyPassManager.h"
 #include "llvm/IR/PassManager.h"
+#include "llvm/IR/PassInstrumentation.h"
 #include "llvm/IR/ValueHandle.h"
 #include "llvm/Pass.h"
