+++ polly/lib/Support/DumpFunctionPass.cpp
@@ -12,6 +12,7 @@
 
 #include "polly/Support/DumpFunctionPass.h"
 #include "llvm/IR/Module.h"
+#include "llvm/IR/PassInstrumentation.h"
 #include "llvm/Pass.h"
 #include "llvm/Support/Debug.h"
 #include "llvm/Support/FileSystem.h"
